Big rewrite, part one
This rewrite fixes some fundamental design mistakes in this module,
making it less error-prone to use and much more maintainable

Properties:
Properties are no longer added at class initialization time, but class
definition time where possible. Where not possible, they will be added
to the classes when the first instance of the class is created. This
avoids repeated introspection

Private bus:
To avoid interfering with any later mainloop, we use a private dbus
connection for the above, which gets disconnected after defining all
classes.

Methods:
Instead of thinking everything is a D-Bus method, we now use
introspection data to define proper methods with the right signatures.

NetworkManager restarts:
auto_reconnect was a bodge and has been ripped out, in favor of a better
way of detecting restarts and recreating proxy objects. It still doesn't
handle objects that get renamed by a restart properly, to be fixed in a
followup commit.

Signals:
The signal handling was not necessarily flawed but also suffered from
the restart problem and could do with a better interface. The signal
handling has been ripped out for now, to be replaced with a better
implementation in a followup commit.

Multiple interfaces:
Device.* and Connection.VPN objects implement multiple interfaces, this
is now properly supported and the SpecificDevice method is no longer
needed.

Documentation:
Actually link to the individual interfaces

Sideeffects:
To do the new definition magic properly, we need to use six; so let's
use it for other python 3 compatibility things too.